added multistory outer wall
Previously, the outer wall was one floor and rendered with as 3 parts,
the outer courtyard building arc, and then two square overpass parts.

For this edit, I deleted the outer courtyard building, redrew the outer
wall arc, and then brought down an upper floor from the interior part of
the building to create a base floor that was in two pieces.  I then deleted
the overpass parts and made them an extension of the building solid.

Because I didn't set the arc side option during the initial drawings, the
outer wall and middle building section do no align perfectly.  This will
need to be addressed in another commit.